requirements.txt
praw==7.6.1
google-cloud-storage
google-cloud-logging
python-dotenv